restoring checkout of update branch to be a new branch
previously we had just git checkout -b <branch> and this
worked because the branch was not fetched on the github
actions host - adding the OR to checkout an existing branch
seems to have triggered an error. All I am doing here
is restoring those lines in both recipes to what they used
to be.

Signed-off-by: vsoch <vsochat@stanford.edu>